+++ src/Classifiers/LogisticRegression.php
@@ -63,7 +63,7 @@ class LogisticRegression implements Estimator, Learner, Online, Probabilistic, V
     protected $optimizer;
 
     /**
-     * The amount of L2 regularization to apply to the parameters of the network.
+     * The strength of the L2 regularization penalty.
      *
      * @var float
      */