+++ src/Classifiers/SoftmaxClassifier.php
@@ -61,7 +61,7 @@ class SoftmaxClassifier implements Estimator, Learner, Online, Probabilistic, Ve
     protected $optimizer;
 
     /**
-     * The amount of L2 regularization to apply to the parameters of the network.
+     * The strength of the L2 regularization penalty.
      *
      * @var float
      */