+++ src/NeuralNet/Layers/Binary.php
@@ -236,7 +236,7 @@ class Binary implements Output
     }
 
     /**
-     * Calculate the gradients for each output neuron and update.
+     * Compute the gradient and loss at the output.
      *
      * @param string[] $labels
      * @param \Rubix\ML\NeuralNet\Optimizers\Optimizer $optimizer