+++ src/Regressors/Adaline.php
@@ -63,7 +63,7 @@ class Adaline implements Estimator, Learner, Online, Verbose, Persistable
     protected $optimizer;
 
     /**
-     * The amount of L2 regularization to apply to the parameters of the network.
+     * The strength of the L2 regularization penalty.
      *
      * @var float
      */